moved rtpArenaSize to FarmingWorldConfig and small API(Impl) cleanup
