+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>at.srsyntax</groupId>
     <artifactId>farmingworld</artifactId>
-    <version>1.4</version>
+    <version>dev-1.5</version>
 
     <repositories>
         <repository>
