+++ src/main/java/at/srsyntax/farmingworld/APIImpl.java
@@ -5,9 +5,9 @@ import at.srsyntax.farmingworld.api.FarmingWorld;
 import at.srsyntax.farmingworld.api.event.DeleteFarmingWorldEvent;
 import at.srsyntax.farmingworld.api.event.GenerateNewFarmingWorldEvent;
 import at.srsyntax.farmingworld.config.FarmingWorldConfig;
+import at.srsyntax.farmingworld.util.LocationRandomizer;
 import lombok.AllArgsConstructor;
 import org.bukkit.*;
-import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
 import org.jetbrains.annotations.NotNull;
@@ -20,7 +20,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
-import java.util.concurrent.ThreadLocalRandom;
 
 /*
  * MIT License
@@ -202,66 +201,8 @@ public final class APIImpl implements API {
   }
 
   @Override
-  public void randomTeleport(Player player, World world) {
-    player.teleport(randomLocation(world));
-  }
-
-  private Location randomLocation(World world) {
-    final boolean nether = world.getEnvironment() == World.Environment.NETHER;
-    int x, y, z, size = plugin.getPluginConfig().getRtpArenaSize();
-    final Location spawn = world.getSpawnLocation();
-
-    do {
-      x = random(spawn.getBlockX(), size);
-      z = random(spawn.getBlockZ(), size);
-      y = world.getHighestBlockYAt(x, z);
-
-
-      if (nether)
-        y = getYInNether(world, x, y, z);
-      else if (y != 0)
-        y++;
-
-      if (!isYValid(world.getBlockAt(x, y-1, z).getType()))
-        y = 0;
-
-    } while (y == 0);
-
-    return new Location(world, x, y, z, spawn.getYaw(), spawn.getPitch());
-  }
-
-  private int getYInNether(World world, int x, int y, int z) {
-    while (y != 0) {
-      y = findBlockAtY(world, x, y, z, true);
-
-      y--;
-      final Block block = world.getBlockAt(x, y, z);
-      if (!block.getType().isAir())
-        continue;
-
-      y = findBlockAtY(world, x, y, z, false) + 1;
-      break;
-    }
-    return y;
-  }
-
-  private int findBlockAtY(World world, int x, int y, int z, boolean air) {
-    Block block;
-    do {
-      block = world.getBlockAt(x, y, z);
-      if (block.getType().isAir() == air) break;
-      y--;
-    } while (y > 0);
-    return y;
-  }
-
-  private boolean isYValid(Material material) {
-    return !this.plugin.getPluginConfig().getSpawnBlockBlacklist().contains(material);
-  }
-
-  private int random(int current, int size) {
-    size/=2;
-    return ThreadLocalRandom.current().nextInt(current - size, current + size);
+  public void randomTeleport(Player player, FarmingWorld farmingWorld) {
+    player.teleport(new LocationRandomizer(this.plugin.getPluginConfig().getSpawnBlockBlacklist(), farmingWorld).random());
   }
 
   public String getDate(long date) {
